backend/requirements-macos.txt
# macOS-optimized requirements for Apple Silicon
# Use this file for macOS installations: pip install -r requirements-macos.txt

attrs>=24.3.0
blinker>=1.9.0
certifi>=2024.12.14
charset-normalizer>=3.4.1
click>=8.1.8
coloredlogs>=15.0.1
contourpy>=1.3.1
cycler>=0.12.1
dnspython>=2.7.0
filelock>=3.16.1
Flask>=3.1.0
flatbuffers>=24.12.23
fonttools>=4.55.3
fsspec>=2024.12.0
huggingface-hub>=0.27.0
humanfriendly>=10.0
idna>=3.10
imageio>=2.36.1
itsdangerous>=2.2.0
Jinja2>=3.1.5
jsonschema>=4.23.0
jsonschema-specifications>=2024.10.1
kiwisolver>=1.4.8
lazy_loader>=0.4
llvmlite>=0.43.0
MarkupSafe>=3.0.2
matplotlib>=3.10.0
mpmath>=1.3.0
networkx>=3.4.2
numba>=0.60.0
numpy>=1.26.0,<2.0.0
opencv-python>=4.10.0.84
packaging>=24.2
pandas>=2.2.3
pillow>=11.1.0
platformdirs>=4.3.6
pooch>=1.8.2
protobuf>=5.29.3
psutil>=6.1.1
py-cpuinfo>=9.0.0
pymongo>=4.10.1
pyparsing>=3.2.1
python-dotenv>=1.0.0
python-dateutil>=2.9.0.post0
pytz>=2024.2
PyYAML>=6.0.2
referencing>=0.35.1
regex>=2024.11.6
requests>=2.32.3
rpds-py>=0.22.3
safetensors>=0.5.0
scikit-image>=0.25.0
scipy>=1.14.1
seaborn>=0.13.2
setuptools>=75.6.0
six>=1.17.0
sympy>=1.13.1
tifffile>=2024.12.12
tokenizers>=0.21.0
# PyTorch with MPS support for Apple Silicon
torch>=2.5.1
torchvision>=0.20.1
tqdm>=4.67.1
transformers>=4.47.1
typing_extensions>=4.12.2
tzdata>=2024.2
ultralytics>=8.3.57
ultralytics-thop>=2.0.13
urllib3>=2.3.0
Werkzeug>=3.1.3